pyproject.toml
[project]
name = "stoatix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "TylerFlar", email = "tulikkovulpus@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
]

[project.scripts]
stoatix = "stoatix.cli:app"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
